add review fixes
