+++ src/features/exports/components/ExportDetailsSidePanel/ExportDetails.tsx
@@ -39,6 +39,7 @@ const ExportDetails: FC<ExportDetailsProps> = ({ job }) => {
     try {
       const result = await onDownload(job);
       if (result) {
+        popSidePathUntilClear();
         notify.success({
           title: "TSV download started",
           message: `${job.name} has been downloaded and removed from the export list.`,
@@ -52,6 +53,7 @@ const ExportDetails: FC<ExportDetailsProps> = ({ job }) => {
   const handleConfirmCancel = async () => {
     try {
       await onCancel(job.id);
+      popSidePathUntilClear();
       notify.success({
         title: "TSV generation cancelled",
         message: `${job.name} has been cancelled.`,
@@ -180,6 +182,7 @@ const ExportDetails: FC<ExportDetailsProps> = ({ job }) => {
                   <ProgressBar
                     progress={job.progress}
                     secondsRemaining={job.estimated_seconds_remaining ?? null}
+                    label={`${job.name} export progress`}
                     fullWidth
                   />
                 ) : (
