+++ src/features/exports/components/ExportsList/ExportsList.tsx
@@ -63,6 +63,7 @@ const ExportsList: FC<ExportsListProps> = ({ exportJobs }) => {
               <ProgressBar
                 progress={job.progress}
                 secondsRemaining={job.estimated_seconds_remaining ?? null}
+                label={`${job.name} export progress`}
               />
             );
           }
