+++ src/features/exports/index.ts
@@ -2,7 +2,7 @@ export { default as ExportsContainer } from "./components/ExportsContainer";
 export { default as ExportDetailsSidePanel } from "./components/ExportDetailsSidePanel";
 export { default as ExportForm } from "./components/ExportForm";
 export { default as SortableFieldList } from "./components/SortableFieldList";
-export { buildExportQuery, getExportScope } from "./helpers";
+export { buildExportQuery, getExportScope, getExportTitle } from "./helpers";
 export type { ExportJob } from "./types/ExportJob";
 export type {
   ExportField,
